README.md
This blackjack code is a simple blackjack game based on randomization. User is allowed to choose accords as long as his score remains lesser than 21. Once the user stops drawing cards, the computer is draws card up to the score of the user. In the end the scores are compared and seen if any of the players has a blackjack!